move utils to right place and add requirements
