+++ backend/requirements/default.txt
@@ -101,3 +101,5 @@ prometheus_client==0.21.0
 fastapi-limiter==0.1.6
 prometheus_fastapi_instrumentator==7.1.0
 sendgrid==6.11.0
+voyageai==0.2.3
+cohere==5.6.1
